1.2.0
WPCS validate update

+++ includes/wpgen/wpgen-handler.php
@@ -1,24 +1,25 @@
 <?php
 /**
- * load more
+ * Ajax wpgen
  *
- * @package fertilizer
+ * @package wpgen
  */
 
-if ( !defined( 'ABSPATH' ) ) {
-	exit; // Exit if accessed directly
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
 }
 
 add_action( 'wp_ajax_ajax_wpgen', 'ajax_wpgen_callback' );
 add_action( 'wp_ajax_nopriv_ajax_wpgen', 'ajax_wpgen_callback' );
 
-// Обработчик
+/**
+ * Ajax callback wpgen form.
+ */
 function ajax_wpgen_callback() {
 
-	array_walk_recursive( $_POST, 'sanitize_form_field' ); // очистка массива
-	parse_str( $_POST['content'], $data ); // Создаём массив который содержит значения полей заполненной формы
+	$_POST = wp_unslash( array_map( 'esc_attr', $_POST ) ); // Очистка массива.
 
-	if ( $_POST['type'] === 'reset' ) {
+	if ( isset( $_POST['type'] ) && $_POST['type'] === 'reset' ) {
 
 		delete_option( 'wpgen_options' );
 		delete_option( 'root_options' );
@@ -27,31 +28,35 @@ function ajax_wpgen_callback() {
 
 	} else {
 
-		$customizer_options = get_option( 'wpgen_options', false );
+		if ( ! empty( $_POST['content'] ) ) {
+			parse_str( wp_unslash( $_POST['content'] ), $data ); // Создаём массив который содержит значения полей заполненной формы.
 
-		if ( !$customizer_options ) {
-			$customizer_options = array();
-		}
-		
-		foreach ( $data as $key => $value ) {
-			if ( stripos( $key, 'customizer' ) !== false ) {
+			$customizer_options = get_option( 'wpgen_options', false );
 
-				$option = str_replace( 'customizer-', '', $key );
-				$option = str_replace( '-', '_', $option );
-
-				$customizer_options[$option] = $value;
-				unset( $data[$key] );
-
+			if ( ! $customizer_options ) {
+				$customizer_options = array();
 			}
+
+			foreach ( $data as $key => $value ) {
+				if ( stripos( $key, 'customizer' ) !== false ) {
+
+					$option = str_replace( 'customizer-', '', $key );
+					$option = str_replace( '-', '_', $option );
+
+					$customizer_options[ $option ] = $value;
+					unset( $data[ $key ] );
+
+				}
+			}
+
+			update_option( 'wpgen_options', $customizer_options, true );
+			update_option( 'root_options', $data, true );
+
+			return wp_send_json_success( $data );
 		}
-
-		update_option( 'wpgen_options', $customizer_options, true );
-		update_option( 'root_options', $data, true );
-
-		return wp_send_json_success( $data );
-
 	}
 
+	return wp_send_json_error();
 
 	wp_die();
